Backend/Genai/requirements.txt
google-genai
python-dotenv
absl-py
flask_cors
flask